% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_codebook.R
\docType{data}
\name{student_codebook}
\alias{student_codebook}
\title{Student Survey Codebook}
\format{
## student_codebook
A dataframe with 159 rows and 6 columns:
\describe{
\item{Variable}(Names of the columns in the student data)
\item{Question}(The survey question)
\item{response_options}(A semi-colon separated character vector of all possible
response options for given Variable/Question)
\item{section}(Section of the survey)
\item{subset}(Whether question was for all, Elementary, or Middle/High students)
\item{type}(Variable type)
}
}
\usage{
student_codebook
}
\description{
The codebook to the 2024-2025 Student Health Survey
}
\keyword{datasets}
