% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_codebook_factors.R
\docType{data}
\name{student_codebook_factors}
\alias{student_codebook_factors}
\title{Factor Levels for the Student Survey}
\format{
A list of length 59. Each element is a character vector of the response options
for the associated question/variable, indicated by the element name.
}
\usage{
student_codebook_factors
}
\description{
Factor Levels for the Student Survey
}
\keyword{datasets}
